add backup and clean up in profile overlay, fix mood card

+++ lib/databases/db_helper.dart
@@ -53,18 +53,19 @@ class DBHelper {
   // FUNGSI RESET DELETE
   Future<int> deleteAllFood() async {
     final db = await database;
+
     return await db.delete('food_diary');
   }
 
   // FUNGSI DELETE DATA LAMA
   Future<int> deleteOldData(int days) async {
+    if (days <= 0) return 0;
+
     final db = await database;
-    // Menghitung tanggal batas (misal: hari ini - 30 hari)
     DateTime limitDate = DateTime.now().subtract(Duration(days: days));
 
     return await db.delete(
       'food_diary',
-      // Gunakan parameter bernama 'where' dan 'whereArgs'
       where: 'time < ?',
       whereArgs: [limitDate.toIso8601String()],
     );
